new-system/requirements.txt
# ECS - 多Agent真涌现系统
# Python依赖包

# ============================================================
# 核心依赖
# ============================================================

# Anthropic API (Claude)
anthropic>=0.40.0

# OpenAI API (可选)
openai>=1.12.0

# ============================================================
# 数据处理
# ============================================================

# NumPy - 数值计算
numpy>=1.24.0

# SciPy - 科学计算
scipy>=1.10.0

# Scikit-learn - 机器学习（用于相似度计算等）
scikit-learn>=1.3.0

# ============================================================
# 配置管理
# ============================================================

# YAML配置文件支持
pyyaml>=6.0

# Python-dotenv - 环境变量管理
python-dotenv>=1.0.0

# ============================================================
# 日志和调试
# ============================================================

# 日志增强
colorlog>=6.7.0

# 进度条
tqdm>=4.65.0

# ============================================================
# 可选依赖
# ============================================================

# 网络图可视化
networkx>=3.0

# 数据可视化
matplotlib>=3.8.0

# HTTP客户端（用于API调用）
httpx>=0.24.0

# 类型检查
typing-extensions>=4.8.0

# ============================================================
# ECS 3.0 新增依赖
# ============================================================

# 重试机制
tenacity>=8.2.0
